✨ (encrypted-field)
Add EncryptedField model to model encryption and decryption of fields.
If another model in our system needs to encrypt some of its data, it
should declare a `belongs_to` relation to EncryptedField and define
custom setters and getters that delegate to the underlying
EncryptedField. This a) makes it easier to provide precise encryption
of data throughout our application without repeated code and b) makes it
easy to rotate our data encryption keys