✨ (AuthorizationToken): Add #expired? method to model
Note that `#expired?` will return `true` if the expiration time is
in the past by even 1 microsecond. This is reflected in tests by virtue
of the delay between setting the expiration time and the actual
execution. We can ease this up to by-second comparison if needed, by
applying `to_i` on the datetimes in the method (to_i gives a Unix time).